tegra-boot-tools: Switch to 3 boot attempts before falling back to the old boot slots on Xaviers
The old boot slots contain the kernel dtb (which in turn
contains the kernel cmdline with the current rootfs label),
the old kernel, old cboot, etc.

We add a patch which modifies thegra-boot-tools
to lower the boot attempts to 3 and also add
a drop-in for the rollback-altboot service to
mark the current slot as OK to boot from,
since there was no kernel crash so far.

Changelog-entry: tegra-boot-tools: Switch to 3 boot attempts before falling back to the old boot slots on Xaviers
Signed-off-by: Alexandru Costache <alexandru@balena.io>